1) Wrench can be picked up, but it doesn't end up in inventory, 2) Wrench minigame can be entered from level 1, but there's a bug with the wrench so it isn't playable
