+++ Assets/Scripts/WrenchMiniGame/GasShutDown.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 74678bc9810703f43ae0ad04eba60605
+guid: 2fe367b7d8a15f149a38a692fdf8a2bc
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
