+++ ProjectSettings/EditorBuildSettings.asset
@@ -20,4 +20,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Levels/Level 1.unity
     guid: 62705fe3445d94c49a618cc70b00c28c
+  - enabled: 1
+    path: Assets/Scenes/MiniGames/Gas_Value_MiniGame.unity
+    guid: 0e868c106ece0f742a5b6c17c7ef8a61
   m_configObjects: {}
